pyproject.toml
[tool.poetry]
name = "password-manager-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
pm = "password_manager_python.main:app"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.15.1"
pymongo = "^4.10.1"
pynacl = "^1.5.0"
pyinputplus = "^0.2.12"
pyperclip = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
